remove the auto pub get behavior
