+++ lib/src/model/package_builder.dart
@@ -59,11 +59,9 @@ class PubPackageBuilder implements PackageBuilder {
     if (!config.sdkDocs) {
       if (config.topLevelPackageMeta.requiresFlutter &&
           config.flutterRoot == null) {
+        // TODO(devoncarew): We may no longer need to emit this error.
         throw DartdocOptionError(
             'Top level package requires Flutter but FLUTTER_ROOT environment variable not set');
-      }
-      if (config.topLevelPackageMeta.needsPubGet) {
-        config.topLevelPackageMeta.runPubGet(config.flutterRoot);
       }
     }
