+++ test/src/utils.dart
@@ -8,10 +8,12 @@ import 'dart:io';
 
 import 'package:analyzer/file_system/file_system.dart';
 import 'package:analyzer/file_system/memory_file_system.dart';
+import 'package:analyzer/file_system/physical_file_system.dart';
 import 'package:analyzer/src/dart/sdk/sdk.dart';
 import 'package:analyzer/src/test_utilities/mock_sdk.dart';
 import 'package:dartdoc/options.dart';
 import 'package:dartdoc/src/dartdoc_options.dart';
+import 'package:dartdoc/src/failure.dart';
 import 'package:dartdoc/src/generator/generator.dart';
 import 'package:dartdoc/src/markdown_processor.dart';
 import 'package:dartdoc/src/matching_link_result.dart';
@@ -20,6 +22,7 @@ import 'package:dartdoc/src/model/package_graph.dart';
 import 'package:dartdoc/src/package_config_provider.dart';
 import 'package:dartdoc/src/package_meta.dart';
 import 'package:dartdoc/src/warnings.dart';
+import 'package:path/path.dart' as p;
 import 'package:pub_semver/pub_semver.dart';
 
 /// The number of public libraries in testing/test_package, minus 2 for
@@ -64,6 +67,25 @@ Future<DartdocGeneratorOptionContext> generatorContextFromArgv(
       optionSet, pubPackageMetaProvider.resourceProvider);
 }
 
+void runPubGet(String dirPath) {
+  var binPath = p.join(p.dirname(Platform.resolvedExecutable), 'dart');
+  if (Platform.isWindows) {
+    binPath += '.exe';
+  }
+  var result = Process.runSync(
+    binPath,
+    ['pub', 'get'],
+    workingDirectory: dirPath,
+  );
+
+  if (result.exitCode != 0) {
+    var buf = StringBuffer();
+    buf.writeln('${result.stdout}');
+    buf.writeln('${result.stderr}');
+    throw DartdocFailure('pub get failed: ${buf.toString().trim()}');
+  }
+}
+
 Future<PackageGraph> bootBasicPackage(
   String dirPath,
   PackageMetaProvider packageMetaProvider,
@@ -73,6 +95,9 @@ Future<PackageGraph> bootBasicPackage(
   bool skipUnreachableSdkLibraries = true,
 }) async {
   var resourceProvider = packageMetaProvider.resourceProvider;
+  if (resourceProvider == PhysicalResourceProvider.INSTANCE) {
+    runPubGet(dirPath);
+  }
   var dir = resourceProvider.getFolder(resourceProvider.pathContext
       .absolute(resourceProvider.pathContext.normalize(dirPath)));
   return PubPackageBuilder(
